+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [[new object] DayOfMonth with essential behavior (#14)](https://github.com/ngarbezza/oow/issues/14)
 - [[extension] Function#value(...args)](https://github.com/ngarbezza/oow/issues/25)
 - [[extension] String#reverse()](https://github.com/ngarbezza/oow/issues/18)
+- [[extension] Set#remove(obj) and Array#remove(obj)](https://github.com/ngarbezza/oow/issues/22)
 
 ## [1.8.0] - 2019-10-10